+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 #### 0.3.1
 
+- The brightness filter's strengt is now maximum 95%
+
 #### 0.3
 - Renamed /migraineedge to /migraineframe (the old one still works)
 - Renamed /migrainedark to /migrainebrightness (the old one still works)